Simple-RAG-app/requirements.txt
streamlit>=1.26.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
torch>=2.1.0
numpy>=1.24.0
openai>=1.0.0
pinecone-client>=2.2.0
